+++ build-tools-internal/src/main/java/org/elasticsearch/gradle/internal/transport/TransportVersionManagementPlugin.java
@@ -43,7 +43,7 @@ public class TransportVersionManagementPlugin implements Plugin<Project> {
         project.getArtifacts().add(tvReferencesConfig.getName(), collectTask);
 
         var validateTask = project.getTasks()
-            .register("validateTransportVersionReferences", ValidateTransportVersionDefinitionsTask.class, t -> {
+            .register("validateTransportVersionReferences", ValidateTransportVersionReferencesTask.class, t -> {
                 t.setGroup("Transport Versions");
                 t.setDescription("Validates that all TransportVersion references used in the project have an associated definition file");
                 Directory definitionsDir = getDefinitionsDirectory(getResourcesDirectory(project));
